refactor: split TimeTrackingService
