+++ Chrono-backend/src/main/java/com/chrono/chrono/services/TimeTrackingService.java
@@ -244,102 +244,6 @@ public class TimeTrackingService {
         throw new IllegalStateException("createPunchInSeparateTx: Unreachable, max deadlock retries exceeded");
     }
 
-    /**
-     * Hilfsmethode zum Anlegen eines PunchOrder (#4 etc.) im AutoPunchOut.
-     */
-    private TimeTracking createPunchForUser(User user, int punchOrder, LocalDateTime stampTime) {
-        logger.debug("createPunchForUser: user={}, punchOrder={}, stampTime={}", user.getUsername(), punchOrder, stampTime);
-
-        TimeTracking newTt = createPunchInSeparateTx(user, punchOrder, stampTime);
-        if (punchOrder == 4) {
-            newTt.setEndTime(stampTime);
-            newTt.setWorkEnd(stampTime.toLocalTime());
-            timeTrackingRepository.save(newTt);
-            logger.debug("createPunchForUser: #4 => EndTime gesetzt");
-        }
-        return newTt;
-    }
-
-    // -------------------------------------------------------------------------
-    // AUTO-PUNCH-OUT
-    // -------------------------------------------------------------------------
-    @Scheduled(cron = "0 20 23 * * *", zone = "Europe/Zurich")
-    @Transactional
-    public void autoPunchOut() {
-        LocalDateTime now = LocalDateTime.now();
-        LocalDate today = now.toLocalDate();
-        logger.info("autoPunchOut gestartet ({} {})", today, now.toLocalTime());
-
-        LocalDateTime startOfDay = today.atStartOfDay();
-        List<TimeTracking> openPunches =
-                timeTrackingRepository.findByEndTimeIsNullAndStartTimeBetween(startOfDay, now);
-
-        // gruppieren
-        Map<User, List<TimeTracking>> byUser = openPunches.stream()
-                .collect(Collectors.groupingBy(TimeTracking::getUser));
-
-        byUser.forEach((user, punches) -> completeMissingPunchesForUser(user, punches, now));
-
-        logger.info("✅ AutoPunchOut beendet – {}", today);
-    }
-
-    /**
-     * Sorgt dafür, dass pro Tag (isPercentage => 2 Stempel, sonst 4) sauber abgeschlossen wird.
-     */
-    private void completeMissingPunchesForUser(User user, List<TimeTracking> openPunches, LocalDateTime punchOutTime) {
-        logger.debug("completeMissingPunchesForUser aufgerufen (user={}, #openPunches={})",
-                user.getUsername(), openPunches.size());
-
-        openPunches.sort(Comparator.comparingInt(tt ->
-                Optional.ofNullable(tt.getPunchOrder()).orElse(0)));
-        TimeTracking last = openPunches.get(openPunches.size()-1);
-        int lastOrder = Optional.ofNullable(last.getPunchOrder()).orElse(0);
-
-        // Prüfe, ob #4 schon existiert
-        Optional<TimeTracking> existing4 = timeTrackingRepository
-                .findFirstByUserAndDailyDateAndPunchOrder(user, punchOutTime.toLocalDate(), 4);
-        if (existing4.isPresent()) {
-            logger.warn("AutoPunchOut {}: PunchOrder #4 existiert schon, skip", user.getUsername());
-            return;
-        }
-
-        switch (lastOrder) {
-            case 1 -> {
-                // Nur WORK_START => #4 => 2 Stempel
-                createPunchForUser(user, 4, punchOutTime);
-                logger.info("completeMissingPunchesForUser: user={}, lastOrder=1 => #4 erzeugt => 2 Stempel", user.getUsername());
-            }
-            case 2 -> {
-                // Pause gestartet, kein BREAK_END
-                if (Boolean.TRUE.equals(user.getIsPercentage())) {
-                    // rename #2 => #4
-                    last.setPunchOrder(4);
-                    last.setEndTime(punchOutTime);
-                    last.setWorkEnd(punchOutTime.toLocalTime());
-                    timeTrackingRepository.save(last);
-                    logger.info("completeMissingPunchesForUser: user={}, lastOrder=2 => rename->#4 (isPercentage)", user.getUsername());
-                } else {
-                    // #3 + #4
-                    createPunchForUser(user, 3, punchOutTime);
-                    createPunchForUser(user, 4, punchOutTime);
-                    logger.info("completeMissingPunchesForUser: user={}, lastOrder=2 => #3+#4 erzeugt => 4 Stempel", user.getUsername());
-                }
-            }
-            case 3 -> {
-                // Pause beendet => #4
-                createPunchForUser(user, 4, punchOutTime);
-                logger.info("completeMissingPunchesForUser: user={}, lastOrder=3 => #4 erzeugt => 4 Stempel", user.getUsername());
-            }
-            case 4 -> {
-                logger.warn("completeMissingPunchesForUser: user={}, #4 existiert, skip", user.getUsername());
-            }
-            default -> {
-                logger.info("completeMissingPunchesForUser: user={}, fallback lastOrder={}, force #4", user.getUsername(), lastOrder);
-                createPunchForUser(user, 4, punchOutTime);
-            }
-        }
-    }
-
     // -------------------------------------------------------------------------
     // TÄGLICHE BALANCE UM 00:05
     // -------------------------------------------------------------------------
@@ -491,103 +395,6 @@ public class TimeTrackingService {
         return convertToResponse(noteEntry);
     }
 
-    @Transactional
-    public String updateDayTimeEntries(String targetUsername,
-                                       String date,
-                                       String workStartStr,
-                                       String breakStartStr,
-                                       String breakEndStr,
-                                       String workEndStr,
-                                       String adminUsername,
-                                       String adminPassword,
-                                       String userPassword) {
-        logger.info("updateDayTimeEntries: target={}, date={}", targetUsername, date);
-
-        User targetUser = userRepository.findByUsername(targetUsername)
-                .orElseThrow(() -> new UserNotFoundException("User not found"));
-
-        checkAdminAndUserPasswords(targetUser, adminUsername, adminPassword, userPassword);
-
-        LocalDate parsedDate = LocalDate.parse(date);
-        LocalTime newWorkStart = LocalTime.parse(workStartStr);
-        LocalTime newWorkEnd   = LocalTime.parse(workEndStr);
-
-        boolean hasBreak = !(breakStartStr.equals("00:00") && breakEndStr.equals("00:00"));
-        LocalTime newBreakStart = hasBreak ? LocalTime.parse(breakStartStr) : null;
-        LocalTime newBreakEnd   = hasBreak ? LocalTime.parse(breakEndStr)   : null;
-
-        LocalDateTime newWorkStartDT  = parsedDate.atTime(newWorkStart);
-        LocalDateTime newWorkEndDT    = parsedDate.atTime(newWorkEnd);
-        LocalDateTime newBreakStartDT = hasBreak ? parsedDate.atTime(newBreakStart) : null;
-        LocalDateTime newBreakEndDT   = hasBreak ? parsedDate.atTime(newBreakEnd)   : null;
-
-        // 1) Alte Einträge löschen
-        int deleted = timeTrackingRepository.deleteByUserAndDailyDate(targetUser, parsedDate);
-        logger.info("updateDayTimeEntries: {} alte Einträge für {} gelöscht.", deleted, date);
-
-        // 2) Neu anlegen: #1, #2, #3, #4
-        // #1
-        TimeTracking ts1 = new TimeTracking();
-        ts1.setUser(targetUser);
-        ts1.setPunchOrder(1);
-        ts1.setDailyDate(parsedDate);
-        ts1.setStartTime(newWorkStartDT);
-        ts1.setWorkStart(newWorkStart);
-        ts1.setCorrected(true);
-        timeTrackingRepository.save(ts1);
-
-        if (hasBreak) {
-            // #2
-            TimeTracking ts2 = new TimeTracking();
-            ts2.setUser(targetUser);
-            ts2.setPunchOrder(2);
-            ts2.setDailyDate(parsedDate);
-            ts2.setStartTime(newBreakStartDT);
-            ts2.setBreakStart(newBreakStart);
-            ts2.setCorrected(true);
-            timeTrackingRepository.save(ts2);
-
-            // #3
-            TimeTracking ts3 = new TimeTracking();
-            ts3.setUser(targetUser);
-            ts3.setPunchOrder(3);
-            ts3.setDailyDate(parsedDate);
-            ts3.setStartTime(newBreakEndDT);
-            ts3.setBreakEnd(newBreakEnd);
-            ts3.setCorrected(true);
-            timeTrackingRepository.save(ts3);
-        }
-
-        // #4
-        TimeTracking ts4 = new TimeTracking();
-        ts4.setUser(targetUser);
-        ts4.setPunchOrder(4);
-        ts4.setDailyDate(parsedDate);
-        ts4.setStartTime(newWorkEndDT);
-        ts4.setEndTime(newWorkEndDT);
-        ts4.setWorkEnd(newWorkEnd);
-        ts4.setCorrected(true);
-        timeTrackingRepository.save(ts4);
-
-        // 3) Balance neu berechnen
-        try {
-            List<String> trackedDates =
-                    timeTrackingRepository.findAllTrackedDateStringsByUser(targetUser.getId());
-
-            int newBalance = 0;
-            for (String dStr : trackedDates) {
-                newBalance += computeDailyWorkDifference(targetUser, dStr);
-            }
-            targetUser.setTrackingBalanceInMinutes(newBalance);
-            userRepository.save(targetUser);
-
-            logger.info("✅ updateDayTimeEntries: Balance für {} neu berechnet: {} min", targetUsername, newBalance);
-        } catch (Exception ex) {
-            logger.warn("⚠️ updateDayTimeEntries: Fehler bei Balance-Refresh: {}", ex.getMessage());
-        }
-
-        return "Day entries updated successfully";
-    }
 
     /**
      * Erzeugt einen Tagesbericht (Liste TimeReportDTO) im Datumsbereich.
@@ -676,113 +483,6 @@ public class TimeTrackingService {
     /**
      * Ergänzend: hole TimeTracking-Einträge für user in Zeitspanne
      */
-    public List<TimeTracking> getTimeTrackingEntriesForUserAndDate(User user, LocalDateTime start, LocalDateTime end) {
-        logger.debug("getTimeTrackingEntriesForUserAndDate: user={}, start={}, end={}",
-                user.getUsername(), start, end);
-        return timeTrackingRepository.findByUserAndStartTimeBetween(user, start, end);
-    }
-
-    /**
-     * Hole alle TimeTracks + user => DTO.
-     */
-    public List<AdminTimeTrackDTO> getAllTimeTracksWithUser() {
-        logger.info("getAllTimeTracksWithUser aufgerufen");
-        List<TimeTracking> all = timeTrackingRepository.findAllWithUser();
-        return all.stream()
-                .map(AdminTimeTrackDTO::new)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Admin kann einen DS direkt editieren (Start/End).
-     */
-    @Transactional
-    public AdminTimeTrackDTO updateTimeTrackEntry(
-            Long id,
-            LocalDateTime newStart,
-            LocalDateTime newEnd,
-            String userPassword,
-            String adminUsername,
-            String adminPassword
-    ) {
-        logger.info("updateTimeTrackEntry: id={}, newStart={}, newEnd={}", id, newStart, newEnd);
-        TimeTracking tt = timeTrackingRepository.findById(id)
-                .orElseThrow(() -> {
-                    logger.warn("TimeTracking ID {} nicht gefunden", id);
-                    return new RuntimeException("Time tracking entry not found");
-                });
-        User targetUser = tt.getUser();
-
-        checkAdminAndUserPasswords(targetUser, adminUsername, adminPassword, userPassword);
-
-        tt.setStartTime(newStart);
-        tt.setEndTime(newEnd);
-        TimeTracking updated = timeTrackingRepository.save(tt);
-
-        logger.info("updateTimeTrackEntry: ID={} aktualisiert (start={}, end={})",
-                updated.getId(), updated.getStartTime(), updated.getEndTime());
-
-        return new AdminTimeTrackDTO(
-                updated.getId(),
-                updated.getUser().getUsername(),
-                updated.getStartTime(),
-                updated.getEndTime(),
-                updated.isCorrected(),
-                updated.getPunchOrder() != null ? updated.getPunchOrder() : 0,
-                updated.getUser().getColor()
-        );
-    }
-
-    // -------------------------------------------------------------------------
-    // PASSWORTCHECK ADMIN + USER
-    // -------------------------------------------------------------------------
-    private void checkAdminAndUserPasswords(User targetUser,
-                                            String adminUsername,
-                                            String adminPassword,
-                                            String userPassword) {
-        if (adminUsername != null && !adminUsername.trim().isEmpty()) {
-            logger.debug("checkAdminAndUserPasswords: adminUsername={}, targetUser={}", adminUsername, targetUser.getUsername());
-
-            User adminUser = userRepository.findByUsername(adminUsername)
-                    .orElseThrow(() -> new RuntimeException("Admin user not found"));
-
-            String storedAdminPwd = adminUser.getAdminPassword();
-            if (storedAdminPwd == null || storedAdminPwd.trim().isEmpty()) {
-                storedAdminPwd = adminUser.getPassword();
-                logger.info("checkAdminAndUserPasswords: Kein separates Admin-Passwort, verwende Login-Passwort.");
-            }
-
-            // Selbst-Bearbeitung?
-            if (targetUser.getUsername().equals(adminUsername)) {
-                if (!adminPassword.equals(userPassword)) {
-                    logger.warn("checkAdminAndUserPasswords: Self-Edit, adminPassword != userPassword");
-                    throw new RuntimeException("For self-edit, admin and user passwords must match");
-                }
-                boolean adminPwdMatch = passwordEncoder.matches(adminPassword, storedAdminPwd);
-                boolean userPwdMatch  = passwordEncoder.matches(userPassword, targetUser.getPassword());
-                if (!adminPwdMatch) {
-                    logger.warn("checkAdminAndUserPasswords: admin password falsch");
-                    throw new RuntimeException("Invalid admin password");
-                }
-                if (!userPwdMatch) {
-                    logger.warn("checkAdminAndUserPasswords: user password falsch (self-edit)");
-                    throw new RuntimeException("Invalid user password");
-                }
-            } else {
-                // Fremd-Bearbeitung
-                boolean adminPwdMatch = passwordEncoder.matches(adminPassword, storedAdminPwd);
-                boolean userPwdMatch  = passwordEncoder.matches(userPassword, targetUser.getPassword());
-                if (!adminPwdMatch) {
-                    logger.warn("checkAdminAndUserPasswords: admin password falsch");
-                    throw new RuntimeException("Invalid admin password");
-                }
-                if (!userPwdMatch) {
-                    logger.warn("checkAdminAndUserPasswords: user password falsch (Fremdedit)");
-                    throw new RuntimeException("Invalid user password");
-                }
-            }
-        }
-    }
 
     // -------------------------------------------------------------------------
     // KONVERTIERUNG -> DTO
